[Intents] Add General, ARC, Web App UMA histograms for Intent Chip
We want to record metrics that show when a link capturing entry point
is available, when the user opens an app using the intent chip/ picker,
and when the user opts to change their settings to remember the
selection. There are separate histograms for ARC apps and Web apps, and
a general histogram for all app types for extra granularity.

Bug: 1304533
Change-Id: I586f13a9418f332fccdc5396b2ba85bede292a5e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3574348
Reviewed-by: Trent Apted <tapted@chromium.org>
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Commit-Queue: Vivian Pao <vpao@google.com>
Cr-Commit-Position: refs/heads/main@{#997518}