ime: Implement System PK Mojo API for IMEs to control quick settings.
Certain IMEs, like Chinese and Japanese, have "quick settings", which
are settings accessed via the input method menu on the shelf. These
settings can be modified in two ways:

A) User presses some shortcut key when using the IME. For example,
   pressing Shift toggles the "Language is Chinese" setting in Chinese
   IMEs

B) User clicks on one of the items in the quick settings menu UI.

Therefore the API is bidirectional: ChromeOS can tell the IME that the
quick settings was changed via the UI, and the IME can tell ChromeOS
that the quick settings was changed via a shortcut.

This was previously implemented for extensions, but not for System PK.
The extension API allows the IME extension to directly change the menu
item strings, but this doesn't work for System PK because the System PK
binary can't deal with string localization. The quick settings strings
must therefore live inside Chromium, and the System PK API will use a
strongly typed representation of the quick settings instead.

The strings added in this CL are NOT new strings: they are migrated
from the internal codebase. To ensure that the translations are
identical, after this CL lands, we will request a One-Time-Request (OTR)
translation to directly import the internal translations to Chromium.

Chinese IMEs have 3 separate boolean settings, so each menu item is
a separate toggle. Japanese IMEs have a single enum setting, so the menu
items is more like a dropdown selection where only one item can be
active at a time.

Bug: b/200613051
Skip-Translation-Screenshots-Check: True
Change-Id: Icc95849a30a18be2e1b697232a7feb448a6615c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3591393
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Keith Lee <keithlee@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Claudio M <claudiomagni@chromium.org>
Reviewed-by: John Palmer <jopalmer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997550}