[Intents] Add metrics for settings changes via Supported Links Infobar
The supported links infobar is part of the link capturing flow, and it
appears when the user has opened a webpage in a webapp that they already
have installed on their device. The infobar asks the user whether they
want to change their settings so that device automatically opens similar
webpages in this selected webapp in the future. This CL adds the metrics
to record whenever the user has accepted the Infobar's suggestion to
change their settings.

Bug: 1304533
Change-Id: I041ed3980482a0915b20d47d537cd699e6164d0b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3613043
Commit-Queue: Vivian Pao <vpao@google.com>
Reviewed-by: Tim Sergeant <tsergeant@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997583}