[FencedFrame] Handle the capture for fenced frames with paint preview
The paint preview should correctly handle the capture and
aggregation of data from fenced frames. This CL fixes the DCHECK
crash when the paint preview's capture is requested to the renderer
for the fenced frame and fixes the issue that removes awaiting
subframes list when the fenced frame gets deleted.

Bug: 1320583
Change-Id: Id72ac3a114a9c74a792ef836619cb33dd76aee63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3612560
Reviewed-by: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Cr-Commit-Position: refs/heads/main@{#997591}