Allow :host::part(foo) to match elements inside shadow roots
This patch is based on WebKit’s fix for the same bug:
https://commits.webkit.org/243873@main

When collecting ::part() rules for an element, we need to consider the
rules from the element’s own tree scope, not just the scopes above, so
we rearrange the loop in StyleResolver::MatchPseudoPartRules.

Such ::part() rules in the element’s own scope must not match unless
preceded by :host (or :host() matching one of its containing shadow
hosts), but we already guard against this in SelectorChecker::CheckOne
thanks to SelectorCheckingContext::treat_shadow_host_as_normal_scope.

When matching the selector to the left of ::part(), the rule could be
in the same scope as the subject iff the selector is :host or :host(),
so we may need to call ParentTreeScope in the kShadowPart case of
SelectorChecker::MatchForRelation.

Bug: 980506
Change-Id: I0693a801461815c9c3a894d41c3ed88aae0ee50c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3610986
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Commit-Queue: Delan Azabani <dazabani@igalia.com>
Cr-Commit-Position: refs/heads/main@{#997652}