exo: support background color for overlay
Delegated lacros needs to pass the background color of a promoted
web canvas overlays to ensure the color of the canvas is correct.

It was thought to use a solid color buffer that could potentially
be used as a piece of information attached to a subsurface that
presents a web canvas, and then sorted out on the exo said. However,
that felt like over-engineering (solid color buffers have different
purpose and are not suitable for usage as a background of a promoted
overlay) and it was decided to pass this information via the
surface-augmenter protocol.

At the moment, web canvas are not promoted as overlays and this
CL is the first step to fix that.

Bug: 1223997
Change-Id: I970604194aeda19cad18b6151eeea1c221b280fe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3609992
Reviewed-by: Peter McNeeley <petermcneeley@chromium.org>
Reviewed-by: Kramer Ge <fangzhoug@chromium.org>
Commit-Queue: Maksim Sisov <msisov@igalia.com>
Cr-Commit-Position: refs/heads/main@{#997659}