Suppress Lacros FRE for Kiosk, Public & Guest sessions
In particular to implement the Guest session check change, this
CL removes support for launching the FRE in the primary profile
from a secondary profile. This was supported for a theoretical
use case, but is not reachable in reality due to the multiple
primary profile checks.

Now if we do the very first run of chrome with a secondary profile
for some reason, the FRE will just be skipped until we open the
primary profile.

Bug: 1300109
Change-Id: Iec094bb96ddc9ed575050df02378885ffd3fad60
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3606798
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Nicolas Dossou-Gbété <dgn@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997662}