Simplify and improve DlpContentManager unittest
This CL is a preparation for increasing code coverage in
DlpContentManager and derived classes. Among other changes:
- Removes checks that require private access to class members, when the
same can be verified indirectly, e.g. HasAnyContentCached().
- Removes checks that appear in several tests and do not bring more
value. E.g. every test for a restriction used to have the same check
that if WebContents are destroyed, the action is allowed. It's enough
to verify it in one test.
- Replaces OnDlpRestrictionCheckedCallback lambda with a mock object,
which simplifies assertions and is more concise.

- Adds a unit test for SCREEN_SHARE restrictions similar to the
existing cases.
  - Includes previously uncovered lines to allow screen sharing of
  content that the user previously bypassed a warning for.


Bug: 1304209,1304213
Change-Id: Ia5d47c7ef00cd022d3726e1390cd4d2ed19dd9dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3610155
Commit-Queue: Aida Zolic <aidazolic@chromium.org>
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997686}