mojo standalone service: add FileDescriptorWatcher instance to main()
It makes sense for the standalone service to bind to the sequence that calls
ServiceProcess::Create(). However, when using
`base::FileDescriptorWatcher::Create()` in that sequence, the code will
typically crash with SIGSEGV because there is no FileDescriptorWatcher instance
on that thread (it uses TLS to cache a pointer).

This updates main() to include a FileDescriptorWatcher instance.

Bug: None
Test: tested local build that had been failing
Change-Id: Id3e8082020a440ed9175b9bb8e427af1df4c11be
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3615109
Reviewed-by: Michael Spang <spang@chromium.org>
Commit-Queue: Gabriel Beddingfield <beddingfield@google.com>
Reviewed-by: Kenneth MacKay <kmackay@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997735}