Rename HeartbeatEvent to ManualTestHeartbeatEvent
Class names should have some signs to show that they are for
testing purpose or not.

Rename HeartbeatEvent* to ManualTestHeartbeatEvent* to make clear that
these classes are for manual test purpose only and should not be
depended upon in production code, even it is compiled into the release
binary.

Bug: b:230794330
Fixed: b:230794330
Change-Id: I845eb68b1a1132c228066c7d19bba5f740c081b0
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3615253
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Hong Xu <xuhong@google.com>
Reviewed-by: Ramin Halavati <rhalavati@chromium.org>
Reviewed-by: Vignesh Shenvi <vshenvi@google.com>
Cr-Commit-Position: refs/heads/main@{#997793}