MPArch:Update a comment in browser_controls_navigation_state_handler.cc
BrowserControlsNavigationStateHandler is observing a WebContents and
changing UI conditions that affect the visibility of top UI such as Tab.
This part code is expected to work only for primary pages as it triggers
a Tab specific implementation. This CL just updates the
auto-added comment for now.

Bug: 1218946
Change-Id: Iff687abdb02c1a1dd209de73065df5b665ccc2f9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3610679
Reviewed-by: John Abd-El-Malek <jam@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Abhijeet Kandalkar <abhijeet@igalia.com>
Cr-Commit-Position: refs/heads/main@{#997813}