full_restore: Do not pass restore info for Lacros windows.
Restore info is set before windows are created, so this is no longer
needed.

This CL does the following:
- Remove lacros_window_handler, which set restore ids when browser windows were created and is no longer needed
- Remove most references to
OnBrowserWindowCreated/OnLacrosBrowserWindowCreated, which also set restore ids and are no longer needed
- Move other logic from On_WindowCreated to OnWindowInitialized

This is part of the effort to support Lacros full restore. The hidden
container and `window_candidates_` will be removed in a follow-up CL.

Test: exo_unittests, browser_tests
Bug: 1291799, 1314763
Change-Id: Id937b708aa5aea9d3d5edbd39e5e0982bf34fc4c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3605753
Reviewed-by: Nancy Wang <nancylingwang@chromium.org>
Commit-Queue: Sophie Wen <sophiewen@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997818}