SharedImageBackingOzone::EndAccess Add fence null checks
SharedImageBackingOzone keeps a map of all the read_fences_ based on
the AccessStream when there is an EndRead. Certain SIRepresentations
such as Dawn, Vulkan, GL call SIBackingOzone::EndAccess to clear the
fence states in the backing on failures OR if the fence is not needed.
In such cases the fence passed is null. This null was added to the
read_fences_ map per AccessStream and later waited upon on the next
SIBackingOzone::BeginAccess call. This CL fixes that issue.

Bug: 1320990
Change-Id: Idae9518744394344df26a71a02321095001aab4e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3616662
Reviewed-by: Vasiliy Telezhnikov <vasilyt@chromium.org>
Commit-Queue: Saifuddin Hitawala <hitawala@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997820}