[PowerMetricsReporter] Make `GetBatteryDischargeDuringInterval` reusable
A future CL will add a new battery metric and another battery discharge
will need to be calculated for another interval. This CL makes
`GetBatteryDischargeDuringInterval` a free function in power_metrics.h
and move the state modification to outside the function so that it can
be reused for the future metric.

In addition, the validation of the interval length was moved out of the
`ReportBatteryHistograms` function so that it can also be reused.

Bug: 1248057
Change-Id: I922d60ffb8ea05e781312e0a577d1bd94d1f422f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3597654
Commit-Queue: Patrick Monette <pmonette@chromium.org>
Reviewed-by: Francois Pierre Doray <fdoray@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997821}