[PaymentRequest] Fix red header for payment handlers
PaymentHandlerWebFlowViewController does color calculations outside
of the typical View system. A refactoring in the Background classes
caused the color to no longer be set when we tried to use it, which
resulted in a default SkRED being returned. As a short-term fix,
forcibly update the Background before trying to use the color.

Bug: 1320311
Change-Id: I80f62bbe16d68fcec9f21d11e1e24c5b4c51cf09
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3615993
Reviewed-by: Allen Bauer <kylixrd@chromium.org>
Commit-Queue: Stephen McGruer <smcgruer@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997861}