Retain drag, autoscroll state if scrollbar is unregistered
Previous behavior was that autoscroll and drag state would be unconditionally removed on DidUnregisterScrollbar, which caused pages with transient changes in overflow behavior or content height to be difficult to scroll.

New behavior is to change the lifecycle of autoscroll and drag state so they are only cleared on OnMouseUp. When scrollbar is unregistered, animations are cancelled and autoscroll/drag state exists in an inactive waiting state until the scrollbar is registered again, in which case scrolling is resumed, or the scroll is cancelled and state cleared on OnMouseUp.


Bug: 1009168

Change-Id: I29e2cab27f6706627d8b316acdd5ce78e781e4c4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3575056
Reviewed-by: Steve Kobes <skobes@chromium.org>
Commit-Queue: Claire Chambers <clchambers@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#997881}