[WebAuthn] Plumb Conditional UI credentials through content layer
(Android only)
When a Conditional UI getAssertion request happens on Android, the list
of credentials has to be provided to the Conditional UI delegate in
order for it to be accessible in autofill. This adds the JNI and
RenderFrameHost methods to allow that to happen.

Bug: 1171985
Change-Id: I8ce35d81e0ee9c6da87ac6022d17fd273df27246
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3534459
Reviewed-by: Bo Liu <boliu@chromium.org>
Commit-Queue: Ken Buchanan <kenrb@chromium.org>
Reviewed-by: Adam Langley <agl@chromium.org>
Reviewed-by: Nina Satragno <nsatragno@chromium.org>
Reviewed-by: Charlie Reis <creis@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997892}