HTTPS-Only Mode: Handle redirects from upgrade HTTPS to HTTP URLs
This CL fixes an issue with HTTPS-Only Mode on iOS when the upgraded
HTTPS URL redirects to an HTTP URL. On the desktop version of the
feature, such a redirect is blocked with the HTTPS-Only Mode
interstitial to prevent the user from landing on an HTTP URL.

This was broken on iOS due to an incorrect assumption that the
URL having an HTTP scheme meant that the navigation shouldn't have the
"upgraded" flag set. This isn't true in case of HTTPS to HTTP
redirects such as the following scenario:
1. User types http://example.test
2. HTTPS-Only Mode upgrades the URL to https://example.test.
3. https://example.test redirects to http://example.test (or any other
HTTP URL)

In this case, the response URL will be HTTP and the pending navigation
will have its IsUpgradedToHttps() flag set to true.

This CL properly handles this case and adds a test for it.

Bug: 1302509
Change-Id: I270630bdba61ddcfdf7c421a7a969bdf11f0bbea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3597667
Reviewed-by: Chris Thompson <cthomp@chromium.org>
Commit-Queue: Mustafa Emre Acer <meacer@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997922}