CloseAll: Add support for undoing desk removal
This CL allows for the desk controller to temporarily save a desk and
its data to a member variable when it is being removed with all of its
windows. This CL also creates a way to restore a desk closed by this
method using a toast. Otherwise, the desk gets disposed when the toast
expires through the toast overlay's destructor. It also adds more unit
tests to validate that removed desks are being cleared when other desks
are being added or removed, as well as whether they are restored or
destroyed when the undo button is pressed or ignored, respectively. This
CL also sets up a new animation for when a desk is restored via the Undo
button.

Bug: 1311314
Change-Id: I07b21aac53f149437fa57a7b9fd79ceee7bec405
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3584270
Commit-Queue: Ben Becker <benbecker@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997979}