[FedCM] Close bubble when tab is detached from its TabStripModel
This CL fixes a couple of things:
* Navigating away is better captured with PrimaryPageChanged()
* Adds a TabStripModelObserver so we can detect when the tab is detached
from the browser window, as currently the bubble will remain in the old
browser window when this occurs since there is no WebContentsObserver
change. When we detect the WebContents is removed from the tab strip, we
close the bubble.

It might be better to have the bubble move with the tab when the tab is
dragged onto a new browser window. However, this requires a larger
change and this intermediate state is acceptable while the more complex
logic is implemented.

Bug: 1296263
Change-Id: I4e57df82f72b61066e218be28ff8fc3ad66f697b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3615978
Reviewed-by: Christian Biesinger <cbiesinger@chromium.org>
Commit-Queue: Nicolás Peña <npm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997981}