Move feed side panel mojo into main.
The Mojo interface for the feed in the side panel will be the interface that Discover Feed content in the chrome-untrusted:// frame will use to communicate with the browser. This interface will provide functionality such as; navigating the frame, opening in a new tab, reporting metrics, etc.

Change-Id: I6a3c12fda9439cf897c82949667408c45c0ee88d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3581291
Reviewed-by: Jian Li <jianli@chromium.org>
Reviewed-by: Ken Buchanan <kenrb@chromium.org>
Commit-Queue: Jonathan Freed <freedjm@chromium.org>
Cr-Commit-Position: refs/heads/main@{#997987}