TSF: remove the workaround for PPAPI Flash
Given that it's been more than a year since Adobe stopped supporting
Flash (on 12/31/2020) and blocked Flash content from running (1/12/2021)
[1], this CL removes the workaround [2] for PPAPI Flash to make the code cleaner and also address a bug caused by the workaround where in EditContext[3], we should return TS_E_NOLAYOUT instead of the caret
bound if the query index is invalid [4] (eg. TSF is querying the bounds
for the 2nd character, and at the moment we only have bounds info for
the 1st character because the author hasn't called
EditContext.UpdateCharacterBounds for the 2nd character. In this
case we should return TS_E_NOLAYOUT to let TSF know the layout info
is not ready)

[1] https://www.adobe.com/products/flashplayer/end-of-life.html
[2] https://chromium-review.googlesource.com/c/chromium/src/+/1603822
[3] https://alexkeng.github.io/edit-context/index.html
[4] https://source.chromium.org/chromium/chromium/src/+/main:content/browser/renderer_host/render_widget_host_view_aura.cc;drc=8ae0d11003e869c96a08346d2d53b8cff9bcc8ef;l=1412

Bug: 963706
Change-Id: Iafa9f45ef1ff02954ae4bb75dc383cf91e20341f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3614988
Reviewed-by: Siye Liu <siliu@microsoft.com>
Commit-Queue: Alex Keng <shihken@microsoft.com>
Reviewed-by: Yohei Yukawa <yukawa@chromium.org>
Cr-Commit-Position: refs/heads/main@{#998141}