Eyedropper module: Use WrapCallbackInScriptScope and update script promise resolver.
In CL 3586744 Reilly proposed to add ExceptionContext to script promise resolver so it can create DOM exception in a single shot.

This CL is a part of chain of updates for replacing DOMException constructor with V8ThrowDOMException and ExceptionState.

Issue: Replace DOMException with ExceptionState/V8ThrowDOMException
Bug: 991544
Change-Id: I957b2e95140ce97c20995e6ac0db97876fd0873b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3577968
Reviewed-by: Yuki Shiino <yukishiino@chromium.org>
Reviewed-by: Kentaro Hara <haraken@chromium.org>
Commit-Queue: Vadym Kargin <vadym.kargin@accenture.com>
Cr-Commit-Position: refs/heads/main@{#998211}