[mparch] Avoid rewriting URLs for Fenced Frames
This CL avoids rewriting URLs for Fenced Frames from
CreateNavigationRequestFromLoadParams() by checking if the frame tree
node is the outermost main frame.
It also extends a parameter with `rewrite_virtual_urls`, which is true
for the outermost main frame, for CreateNavigationEntry() to know if it
needs to rewrite URLs since it also calls RewriteUrlForNavigation().

Bug: 1314749
Change-Id: I269f9903d06c802f629f677ca3f342f6e43ae356
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3602352
Reviewed-by: Charlie Reis <creis@chromium.org>
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Commit-Queue: Julie Jeongeun Kim <jkim@igalia.com>
Cr-Commit-Position: refs/heads/main@{#998217}