[Fenced Frame] Only get favicon urls of a primary main frame
We have a restriction that updates only the favicon URLs for the
primary main frame in WebContents, and the fenced frame's favicon
URLs are not updated. So, the favicon URL should be empty for the
fenced frame like an iframe in ContentCaptureReceiver when the
content capture retrieves them.

Bug: 1319357
Change-Id: I875e7fcfb394f82bb35cf18a866d4096560c9a63
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3600104
Reviewed-by: Kevin McNee <mcnee@chromium.org>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Commit-Queue: Miyoung Shin <myid.shin@igalia.com>
Reviewed-by: Michael Bai <michaelbai@chromium.org>
Cr-Commit-Position: refs/heads/main@{#998223}