lp: Add a method to LanguagePackManager to install base payloads.
We are adding the concept of "Base Payloads" to Language Packs. Base
Payloads are language-agnostic dependencies that need to be installed
in order for a Language Pack to be used. For example, the Language Packs
for handwriting recognition need the handwriting recognition library to
be installed as a Base Payload first.

As an initial step, add an explicit method install the base payload.
In the future, we will modify InstallPack so that it will automatically
install the base payload corresponding to a language pack.

Bug: 1320137
Change-Id: I15f91e447873d8ca9454fe10c296e2fc75c914c9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3611431
Commit-Queue: Darren Shen <shend@chromium.org>
Reviewed-by: Claudio M <claudiomagni@chromium.org>
Cr-Commit-Position: refs/heads/main@{#998240}