crOS: Change fake account emails on tests
Change the domain of the fake account emails on
`inline_login_handler_chromeos_browsertest.cc` from `example.com`
to `gmail.com`.
The reason of this change is a new feature (crbug/1289296) that checks
if the account is enterprise account through an API. Making the domain
a `gmail.com` would prevent calling this API to ask for the type of
user account since `gmail.com` are well known general user accounts.

Bug: 1292366
Change-Id: I18b66aea59db84bd560382b6c6659942205ebbfe
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3606081
Reviewed-by: Anastasiia N <anastasiian@chromium.org>
Commit-Queue: Martin Rodriguez <rodmartin@google.com>
Cr-Commit-Position: refs/heads/main@{#998294}