[Lacros] Improve ProfilePicker window re-use
The SigninManager is now able to re-use an existing ProfilePicker
window.
The existing callbacks are called immediately with error values, and
are replaced by the new callbacks. The picker window is also refocused.

This fixes an issue in SigninManager where a new signin flow could not
be started when one was already in progress.

Bug: 1319416
Change-Id: I98d42dac0186b450369e1f3c831be43b6360c09a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/3608014
Reviewed-by: Nicolas Dossou-Gbété <dgn@chromium.org>
Commit-Queue: David Roger <droger@chromium.org>
Cr-Commit-Position: refs/heads/main@{#998312}