Exclude ContentSettingsPattern port separator from path
After determining the scheme, we currently zoom ahead to the first
occurrence of ':' and treat it as the port separator. What this misses
is that the path might contain a ':'.

This CL changes the parser to zoom ahead to the first of ':' or '/'.

This works because ':' and '/' are forbidden host code points [0], so
there is no danger of reading part of the host as the port.

I am continuing to use the assumption that the path must begin with '/',
but I admittedly don't have a reference for that claim.

[0]: https://url.spec.whatwg.org/#forbidden-host-code-point

Bug: 1196591, 1134913
Change-Id: I161b3109ace23d97d1f00761ad282ccab33e7610
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2875723
Reviewed-by: Andy Paicu <andypaicu@chromium.org>
Commit-Queue: Andy Paicu <andypaicu@chromium.org>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#998317}